remove test to 2.3 and 3.1 module
