+++ paimon-hive/paimon-hive-connector-2.3/pom.xml
@@ -569,6 +569,12 @@ under the License.
             <version>${iceberg.version}</version>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>org.apache.flink</groupId>
+            <artifactId>flink-metrics-dropwizard</artifactId>
+            <version>${iceberg.flink.dropwizard.version}</version>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 
     <build>
