+++ paimon-hive/paimon-hive-connector-2.3/src/test/java/org/apache/paimon/procedure/MigrateIcebergTableProcedure23ITCase.java
@@ -1,0 +1,24 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.paimon.procedure;
+
+import org.apache.paimon.hive.procedure.MigrateIcebergTableProcedureITCase;
+
+/** IT cases for migrating iceberg table to paimon table in hive 2.3. */
+public class MigrateIcebergTableProcedure23ITCase extends MigrateIcebergTableProcedureITCase {}
