+++ paimon-hive/paimon-hive-connector-3.1/pom.xml
@@ -599,6 +599,12 @@ under the License.
             <version>${iceberg.version}</version>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>org.apache.flink</groupId>
+            <artifactId>flink-metrics-dropwizard</artifactId>
+            <version>${iceberg.flink.dropwizard.version}</version>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 
     <build>
