+++ paimon-hive/pom.xml
@@ -130,20 +130,6 @@ under the License.
             <version>${project.version}</version>
             <scope>test</scope>
         </dependency>
-
-        <dependency>
-            <groupId>org.apache.iceberg</groupId>
-            <artifactId>iceberg-flink-${iceberg.flink.version}</artifactId>
-            <version>${iceberg.version}</version>
-            <scope>test</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>org.apache.flink</groupId>
-            <artifactId>flink-metrics-dropwizard</artifactId>
-            <version>${iceberg.flink.dropwizard.version}</version>
-            <scope>test</scope>
-        </dependency>
     </dependencies>
 
 </project>
